Fix RPC index of others when an RPC is skipped
If an RPC is skipped because the API key is missing, this affects the `index` property of RPCs with traces. The RPC with trace ends up with a wrong index and when being used it becomes `undefined`.

The fix keeps track of the RPCs skipped and reduces that number from the index.