Remove buggy the refresh_materialized_view function, raise the pg notice to log level (#2396)
* Remove buggy the refresh_materialized_view function, raise the pg notice to log level

Turns out the original function we wrote was buggy as you can't really UPDATE a materialized view. That's why the refresh failed last night. We can completely get rid of the function and just schedule a direct refresh of the materialized view, and it will take care of the `refreshed_at` columns itself

TODO:
- [ ] Execute fix queries on staging
- [ ] Execute fix queries on prod

* Remove 4byte stats.created_at

* Update sql dump

* Change postgres log level to warning

* Update pg_cron setup and README

* Typo

Co-authored-by: Manuel Wedler <manuel@wedler.dev>

---------

Co-authored-by: Manuel Wedler <manuel@wedler.dev>